+++ bin/mapper.tcl
@@ -7,7 +7,7 @@
 # | | ____ | |(_)| ||  ___  |           (____   _)     _/   /    /  /       _/   /     #
 # | | \_  )| |   | || (   ) |                ) (      /   _/    /  /       /   _/      #
 # | (___) || )   ( || )   ( | Mapper         | |   _ (   (__/\ /  /     _ (   (__/\    #
-# (_______)|/     \||/     \| Client         (_)  (_)\_______/ \_/     (_)\_______/    #
+# (_______)|/     \||/     \| Client         (_)  (_)\_______/ \_/     (_)\_______/ __ #
 #                                                                                      #
 ########################################################################################
 # TODO move needs to move entire animated stack (seems to do the right thing when mapper is restarted)
@@ -17,10 +17,10 @@
 # GMA Mapper Client with background I/O processing.
 #
 # Auto-configure values
-set GMAMapperVersion {4.27.2}     ;# @@##@@
+set GMAMapperVersion {4.27.2_alpha.0}     ;# @@##@@
 set GMAMapperFileFormat {23}        ;# @@##@@
 set GMAMapperProtocol {415}         ;# @@##@@
-set CoreVersionNumber {6.26}            ;# @@##@@
+set CoreVersionNumber {6.27}            ;# @@##@@
 encoding system utf-8
 #---------------------------[CONFIG]-------------------------------------------
 #
@@ -14488,7 +14488,7 @@ proc ConnectToServerByIdx {idx} {
 #
 #*user_key name -> sanitized_name
 #
-# @[00]@| GMA-Mapper 4.27.2
+# @[00]@| GMA-Mapper 4.27.2_alpha.0
 # @[01]@|
 # @[10]@| Overall GMA package Copyright © 1992–2024 by Steven L. Willoughby (AKA MadScienceZone)
 # @[11]@| steve@madscience.zone (previously AKA Software Alchemy),